requirements.txt
boto3
python-dotenv
moviepy
ffmpeg
ffmpeg-python
Pillow